Split model types into separate files
Separate the model package types into individual files:
- task.go: Task, TaskStatus, TaskCommand, Instruction, RunnerEvent
- event.go: Event
- link.go: Link
- log.go: Log
