+++ internal/model/task.go
@@ -1,4 +1,3 @@
-// Package model defines shared types used across the xagent codebase.
 package model
 
 import "time"
@@ -25,15 +24,6 @@ const (
 	TaskCommandStop    TaskCommand = "stop"
 )
 
-// RunnerEventType represents the type of event reported by the runner.
-type RunnerEventType string
-
-const (
-	RunnerEventStarted RunnerEventType = "started"
-	RunnerEventStopped RunnerEventType = "stopped"
-	RunnerEventFailed  RunnerEventType = "failed"
-)
-
 // Instruction represents a task instruction with text and optional source URL.
 type Instruction struct {
 	Text string `json:"text"`
@@ -54,39 +44,19 @@ type Task struct {
 	UpdatedAt    time.Time     `json:"updated_at"`
 }
 
+// RunnerEventType represents the type of event reported by the runner.
+type RunnerEventType string
+
+const (
+	RunnerEventStarted RunnerEventType = "started"
+	RunnerEventStopped RunnerEventType = "stopped"
+	RunnerEventFailed  RunnerEventType = "failed"
+)
+
 // RunnerEvent represents an event from the runner about a task's container.
 type RunnerEvent struct {
 	TaskID    int64
 	Event     RunnerEventType
 	Version   int64
 	Reconcile bool
-}
-
-// Event represents an external event that can trigger task actions.
-type Event struct {
-	ID          int64     `json:"id"`
-	Description string    `json:"description"`
-	Data        string    `json:"data"`
-	URL         string    `json:"url,omitempty"`
-	CreatedAt   time.Time `json:"created_at"`
-}
-
-// Link represents a link between a task and an external resource.
-type Link struct {
-	ID        int64     `json:"id"`
-	TaskID    int64     `json:"task_id"`
-	Relevance string    `json:"relevance"`
-	URL       string    `json:"url"`
-	Title     string    `json:"title"`
-	Notify    bool      `json:"notify"`
-	CreatedAt time.Time `json:"created_at"`
-}
-
-// Log represents a log entry for a task.
-type Log struct {
-	ID        int64     `json:"id"`
-	TaskID    int64     `json:"task_id"`
-	Type      string    `json:"type"`
-	Content   string    `json:"content"`
-	CreatedAt time.Time `json:"created_at"`
 }
